Experiment with alternate editor paradigm
Because `innerText` and `computeVDOMCursor` are faster than expected, clocking in at about 1-5ms (each) in production mode, it may be viable to use a more naive, hybrid approach based on the original code demo the sparked the new editor design.

Instead of reading the entire document on input, it may make sense to read a greedy area of the current block DOM node plus and minus one, that is to say, always extend on block DOM node behind and ahead.

This would conceivably fix any unexpected backspace on line, delete on line problems, while also maintaining the smallest possible footprint, and still making use of `vdom.write`, which is expected to be a fast iterator.

It should be noted that any use of `greedy` needs to be updated per input; `selectionchange` does not fire when a user hits delete, because the cursor doesn’t technically move.

Implementing all of this would theoretically speed up the editor considerably, not flush keys (enabling the React renderer to stay fast), and can still be optimized with changes to `VDOM`.