Saving changes following experimentation
It turns out that the patching technique does not work as expected. The concept was to iterate from the node before the anchor node to the new anchor node, read the affected area, and commit a `greedyWrite`.

The result was that holding ctrl-d starting appending the end character, for example, `# How to build a beautiful bloggg`. It is not obvious why this is happening, but having checked both previous implementations; reading the entire DOM, and working with the anchor node, that that problem did not purport to exist.

Took a closer look at Google’s diffing library https://github.com/google/diff-match-patch. If we read the entire DOM (which is faster than expected), then we could do a text diff. The standing concern is that the time required by `diff-match-patch` is now competing with the 16.67 threshold.