Documenting some observations -- see description
It looks like Android has a selection problem. This problem has also been observed in minimal code demo and the Trix editor. This means that on a low-powered device/environment, Android cannot guarantee the cursor will be positioned correctly. That being said, this issue is not observable in a plain textarea.

The problem essentially is that Android appears to bail from a synchronous call to `selection.addRange`. This basically means that Android cannot, as far as I can tell, ever be considered a stable platform for a WYSIWYG editor. Period.

After doing even more experiments with Notion, etc., it may be that if an input event is in progress we need to do nothing. This works like a mutex except we still let the browser do its thing -- we just don’t get involved until after. This could be prototyped by looking at the key up event, assuming Android respects key up.

A mutex would seem to solve a lot of problems, most specifically allowing a user to press and hold keys (cut, copy, paste, etc. are prevented anyway), still experience as-fast-as-possible rendering speeds, and most importantly, relieve stress from the implementation details of how input is handled.