Discovered one unlikely to be reproduced bug (forward word delete)
Holding delete on a word (e.g. fn-opt-delete) manages to break the editor’s width, which is surprising. Appears to only happen on code blocks -- most likely has to do with `pre`.

Speed does not matter -- a single forward delete word event can cause the bug.

This could be prevented (by looking at `inputType`) with a negligible effect on the editor.