Observations about Firefox contenteditable
Native contenteditable is weirder than expected on Firefox; backspace on the insertion point node seems to delete the current node and set the cursor forwards, which is entirely unexpected

It may worth doing more or less what Trix does; we might want to loosely translate native DOM operations to an internal model and rerender. Right now, we’re not actually rerendering on delete, backspace, which is more error prone and less well behaved.

In Firefox we can’t even assume even backspace to be idempotent; therefore, we may want to reimplement all backspace, delete, etc. methods with a hard should rerender. What we loose in doing this is native backspace handling on special characters like emoji.

All hope is not lost; previous implementations of backspace were Unicode aware to the extent that they can be. We can even iterate on this: we can sniff for the zero-width joiner character (\u200d or 8,205) to properly remove emoji graphemes.

This is preferred to sniffing the previous character and conditionally rerendering (rather than hard rerendering) because a) it solidifies our experience of working with Unicode and emoji characters, which is desirable, and b) it’s consistent with the user’s experience without leaving details up to the host browser.

Note that Firefox still appears to sometimes retrieve the root node instead of a child node, which obviously is inconsistent with our model (this has been solved for in the past with success).