Save changes -- experimenting with mobile-first backspace and delete handlers
Before, we were in effect ignoring proper handling for backspace and delete methods. The previous implementation solved for these inputs via `onKeyDown`, which is more or less a no-op on mobile (Android).

To solve for this, there is a new `inputType` package that helps translate input event types into methods (draws a 1:1 relationship). Also, the same handlers used in `onKeyDown` are now being moved to `onInput`; deferring to native rendering for backspace and delete on emojis still needs to be handled (see `FIXME`s).

If this works, in doing so we free ourselves from micro managing `onKeyDown` events (exceptions are tab and formatting), and theoretically provides higher confidence to our `onInput` resolver. In effect, `onInput` would both be resilient to known and unknown side effects within the given greedy DOM node range.

This line of reasoning and code is attempting to solve for the 90% case of input events without needing to double our efforts in `onKeyDown`.