The superfluous Android selection bug has been identifier
The selection bug as noted in the previous commit described a naked `contentEditable` bug.

When pressing delete on:

hello
<cursor>hello

Android spits out:

helloh<cursor>ello

It turns out that this somehow relates to using semantic HTML elements, as supposed to `<div>`, which is as boring as it gets.

The fix for this bug was identified in the client source code of notion.so; Notion appears to use `div`s everywhere imaginable. Therefore, semantic elements are **not** recommended where possible.