Stage `CodeDemo` to be the new main Editor
As a side note, it doesn’t make sense to over optimize for `onInput`; the pressing concerns are solving for Firefox and Android (mobile), simultaneous to making a functional, well behaved editor.

Because Firefox is almost exclusively desktop based, we **can** optimize for `onKeyDown`, and because Android is more well behaved than Firefox, and because we found the code block bug (where backspace lost `<code>`), we can instead reasonably solve for `onKeyDown` instead.

Android may still be error prone when it comes to inserting a newline, because `onKeyDown` is more or less a no-op. Despite this, Android may very well be more well behaved because we are not relying on semantic elements anymore (for components), which led to superfluous selection bugs.

The added bonuses are that the code will be more compartmentalized and less reactive and more proactive.