Come up with a new concept for reconciling the React DOM with the client DOM based on the Unix timestamp idea. See `UnixExperiment.js` for a MVP demo
The basic concept is to attribute greedy DOM nodes with a last modified at timestamp, like an inverse expiry. This would provide a superficial mechanism to understand when was a component last rendered (because it needed to).

In order to achieve this effect, use `React.memo` so that extraneous attributes set *within* component definitions are not needlessly rerendered -- like using the `Math.random` technique to guarantee a component rerenders.