Save changes
Using Mutation Observers as the onInput handler is a little weak because there are more unknown unknowns that known unknowns, which makes it hard to reason about as a new baseline for how the editor works.

Despite that, there’s a very interesting concept solution for how to lightly resolve the React DOM with the client DOM. In essence, rather than implementing some naive diffing algorithm, it is very possible to consider whether Mutation Observers can do the job. Rather than observe changes to the client DOM, we can observe changes to the React DOM and use that information (which is rich in nature) to incrementally patch the client DOM.

That, fortified with the new awareness of how to model the VDOM and VDOM cursors may be enough to make a fast, usable editor that works everywhere.

The proposed changes (based on early research) to the data model are as follows:
the cursor is extracted either from its coordinates or key and relative position. This *should* be sufficient, even for that status bar because that status bar needs the following: line, column, lines selected, characters selected. The total body count can still be extracted from the plain text data value (which is incrementally updates via nodes.join("\n") -- which is faster than you think).

As for how greedy should follow (which has not been sufficiently prototyped) -- the same principles can apply; when a native selectionchange or onKeyDown event occurs, iterate from the stable start node to the end node. With this information, we know in advance which keys will be affected, which should be enough considering we will most likely reread the affected nodes and commit a multi-node update. We may need some heuristic of, if a key appeared or disappeared, we need to add or delete a node. Details need to be sorted out.