Save changes
If we stop trying to separate React from the client DOM, then there may be some patterns based on prior research that just work.

Basically, we could had reset current node’s ID so as to blow it up per render -- if a rerender is required. This actually *should* if the user cannot arbitrarily create DOM nodes (other than mutate text nodes, because I think React always blows up text nodes), and enter, backspace is handled.

In the case of preventing arbitrarily creating DOM nodes, this theoretically can be handled just by preventing formatting, which we know how to do and can be reasonably supported.

In the case of enter and backspace, we need to be more careful. The good news is we can easily determine if a node was added or dropped based on the range or start and end keys.

If a node was dropped, so far we know how to implicitly handle that, because as long as React doesn’t throw, this is more or less a non-issue except for possibly annoying error message, which we know how to tamper.

The special case is *was a node added*. This we can also determine based the presence of a repeat key, which would be thanks to the browser. In any matter, when the presence of a repeat node is determined, we can read the DOM as is then hard reset the DOM by inserting a cloned document fragment.

The good news is that we know in advance the only nodes we possibly need to clone are the current node and the next node -- it’s not the whole selection. The reason we need both is because the behavior appears to be different whether a user hit enter at the beginning or mid or end paragraph.