Ignoring Firefox for now -- can add Puppeteer later and mimic Chrome tests
Firefox appears to pose two problems: added nodes do not copy the node ID from where they came, and backspace on a node (probably delete, too) actually deletes the next node -- not merges it. This means references are lost and previous and next siblings cannot be access (despite still existing in memory).

One potential solution may be to compute an array of references in getTargetInputRange.