Save changes -- the rich editor may be salvageable!
It looks like syncViews may be one of the root causes -- for now, all DOM nodes are regenerated from scratch for sanity.

Also noticed that arrow down (no selection) at the start of a code block (before, not one) jumps the cursor to to the end of the code block (Gecko/Firefox). This is most likely due to how Firefox handles selection on a data-compound-node.