Found two browser-related bugs (Chrome)
white-space: pre, overflow-wrap: break-word do not work as expected (in a textarea context) with the following:

\tsss …

where s overflows. This breaks because of the tab (works as expected with a space instead). Apparently, textarea handles whitespace *differently* than HTML elements. Wha?

Sometimes undo deletes one character at a time instead of a word (expected). There is no obvious answer to this -- a solution would be a custom implemented undo, redo manager.