Fixed two bugs related to the rich text editor
When deleting a range that spans multiple data nodes, if the cursors are reversed and the start node is an empty node, Gecko/Firefox surprisingly **does not** preserve the data-node; a plain text string node is left in its absence, which leads to problems with KeyNodeIterator and the preserving the selection (cannot reset a cursor without a data node ID).

The second bug is more familiar; even when using onSelect and onPointer*, onKeyDown **still** needs to be able to update the target. Before, there was an emergency update the target in rendering stage, but that’s wrong because it still misses onKeyDown events.