Added discretionary timers to getPosFromRange and getData -- warns if equal to or greater than a global constant (e.g. 5 milliseconds)
Surprisingly, an index-based implementation of getData (versus an allocation-based implementation) is **extremely** fast, performing in less than 5 milliseconds, even for huge documents. Because we’re testing on desktop, this isn’t entirely fair, but it does go to show that the *real* bottlenecks in the perceived application performance are (possibly) React and the parser.

Because the (to be) parser is hand-written and optimized (not regex-based), the parser is not necessarily a pressing concern going forwards.