Added (very) basic handling for compositional input, which was previously broken in Gecko/Firefox
It looks the crux of the problem is that Firefox **cannot** be relied on via onInput to process intl. input. This is a non-issue in Chrome because Chrome or more or less well behaved, but in Firefox, correcting the cursor **before** compositional input is committed (e.g. composition end) leads to cursor bugs which leads to worse problems.

By simple deduping composition end events **and** writing repeat logic for composition end (as if an input handler), the editor appears to be a) stable and b) aware of intl. input *when it matters*. There was a conceptual concern that if a user used the arrow keys or reselected and created input, that the data would be corrupt (pre-committed), but due to the nature of composition end, this seems to be a non-concern.