Documenting a Gecko/Firefox bug
When deleting the period at the right-side of the following:

This is **bold**, this is *italic*, and this is ***bold***. And this is `code`.

Firefox, for whatever reason, jumps all the way to the left-hand side. It’s not clear why this is happening, but most most likely a side effect of either:

- Semantic elements
- font-size, etc. mismatch
- Native or synthetic backspace (not fully synthetic -- we can prevent early but that doesn’t necessarily fix mobile)

This is confirmed to not be observable in Chromium/Chrome.