Refactored collapsed to hasSelection and focused to isFocused
It makes more sense that we name variables based on the non-zero value behavior. For example, when the editor loads, it cannot be assumed the editor is focused, so isFocused is the opposite of the zero value. The same is true with hasSelection. This convention should make some decisions going forward easier.