Reimplemented backspace word (backwards and forwards)
The old implementation was based textarea which is actually undesirable because it’s too greedy. The native contenteditable backspace word is more familiar, the same as in Sublime Text (less sensitive). This is the expected behavior (at least for me). Updated tests and entire test suite passes on all platforms.