Save change and documenting notes
**DO NOT PUSH TO PRODUCTION**

It looks like way may be able to reuse write(...) instead of creating another method e.g. dropBytes(...) because the way write works is conceptually *extremely simple*.

As long as pos1 and pos2 have not void (e.g. .x, .y, and .pos), all we theoretically need to do to reuse write for deletion is decrement and increment pos as needed. This is very elegant and is suggestive of how we may solve for smart smart character replacement because such operations require both deleting and inserting text simultaneously.