Documenting performance testing
It looks like Chrome and Firefox can run as fast as 0-2ms (Chrome has issues with emojis if <2ms) and Safari needs >16.67ms to be reliably tested.

If DELAY=x is not specified, playwright tests at 16.67ms. If DELAY=x is specified, playwright tests at <x>ms. For Chrome and Firefox, it appears we can reliably test at 2ms minimum.