Found the select-all type bug
The problem is that our end iterator can’ iterate *past* the end node. This is similar to insertNode(...) -- insertNode *at the end node* takes null as an argument (because there is no node at the end).

I don’t know the best way to fix this yet but it’s a start.

After this is implemented, assuming the test suite passes, we can push to production.