Documenting a bug
Force-holding redo then undo can lead to exception with setting the range. First, we needed to store copies of body, pos1, etc. because they are reference values (that was to be expected), but the following is less so expected:

{ node: null, offset: 0, pos: 8819, length: 8819 }

Essentially, there is an edge case when pos is in bounds but can’t be set, because it’s at the end. Because of this, it’s reasonable to use a try-catch for inside the scope of shouldRender.