Documenting a bug
I think the problem related to press-hold paste and undo is because more undos are being dispatched before the editor is in sync. This is because in the current design, there is no concept of a mutex. I’m not 100% sure this is the problem, but it seems likely because pressing undo slowly does not seem to cause the problem.