Added support for offline mode and caching
firebase.firestore().enablePersistence() is the same thing as enabling the cache and support for offline mode. This is extremely advantageous because it limits the number of useless reads from the server (which are billable), and is faster, near-instant for offline users. Note that persisted writes are still theoretically written as individual writes rather than as single overwrite (all billable), but this isn’t any different than if the user was online the whole time.